+++ wgpu/src/lib.rs
@@ -1783,7 +1783,7 @@ impl Adapter {
                 Queue {
                     context,
                     id: queue.id().into(),
-                    data: Box::new(()),
+                    data: Box::new(queue),
                 },
             )
         })